test: consolidate route selection fixtures
