+++ tests/runtime/conftest.py
@@ -3,6 +3,10 @@ from collections.abc import Callable
 from pathlib import Path
 
 import pytest
+from jsonschema import Draft202012Validator
+
+
+ROOT = Path(__file__).resolve().parents[2]
 
 
 @pytest.fixture
@@ -49,3 +53,15 @@ def make_task_dir(write_json: Callable[[Path, dict], None]) -> Callable[[Path], 
         return task_dir
 
     return _make_task_dir
+
+
+@pytest.fixture
+def assert_schema_valid() -> Callable[[str, dict], None]:
+    def _load_schema(name: str) -> dict:
+        return json.loads((ROOT / "schemas" / f"{name}.schema.json").read_text(encoding="utf-8"))
+
+    def _assert_schema_valid(name: str, payload: dict) -> None:
+        errors = sorted(Draft202012Validator(_load_schema(name)).iter_errors(payload), key=lambda err: list(err.path))
+        assert not errors, [f"{list(err.path)}: {err.message}" for err in errors]
+
+    return _assert_schema_valid
