+++ tests/runtime/test_route_selection.py
@@ -1,70 +1,23 @@
 import json
+from collections.abc import Callable
 from pathlib import Path
-
-from jsonschema import Draft202012Validator
 
 from forgeflow_runtime.orchestrator import escalate_route
 
 
-ROOT = Path(__file__).resolve().parents[2]
-
-
-def _write_json(path: Path, payload: dict) -> None:
-    path.write_text(json.dumps(payload, indent=2), encoding="utf-8")
-
-
-def _load_schema(name: str) -> dict:
-    return json.loads((ROOT / "schemas" / f"{name}.schema.json").read_text(encoding="utf-8"))
-
-
-def _assert_schema_valid(name: str, payload: dict) -> None:
-    errors = sorted(Draft202012Validator(_load_schema(name)).iter_errors(payload), key=lambda err: list(err.path))
-    assert not errors, [f"{list(err.path)}: {err.message}" for err in errors]
-
-
-def _make_task_dir(tmp_path: Path) -> Path:
-    task_dir = tmp_path / "task"
-    task_dir.mkdir()
-    _write_json(
-        task_dir / "brief.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "objective": "Run a small route",
-            "in_scope": ["runtime"],
-            "out_of_scope": [],
-            "constraints": ["local only"],
-            "acceptance_criteria": ["route works"],
-            "risk_level": "low",
-        },
-    )
-    _write_json(
-        task_dir / "run-state.json",
-        {
-            "schema_version": "0.1",
-            "task_id": "task-001",
-            "current_stage": "clarify",
-            "status": "in_progress",
-            "completed_gates": ["clarification_complete"],
-            "failed_gates": [],
-            "retries": {},
-            "current_task_id": "",
-            "spec_review_approved": False,
-            "quality_review_approved": False,
-        },
-    )
-    return task_dir
-
-
-def test_escalate_route_switches_to_large_high_risk(tmp_path: Path) -> None:
-    task_dir = _make_task_dir(tmp_path)
+def test_escalate_route_switches_to_large_high_risk(
+    tmp_path: Path,
+    make_task_dir: Callable[[Path], Path],
+    assert_schema_valid: Callable[[str, dict], None],
+) -> None:
+    task_dir = make_task_dir(tmp_path)
 
     state = escalate_route(task_dir=task_dir, from_route="small")
 
     assert state["status"] == "blocked"
     assert state["current_stage"] == "clarify"
     checkpoint = json.loads((task_dir / "checkpoint.json").read_text(encoding="utf-8"))
-    _assert_schema_valid("checkpoint", checkpoint)
+    assert_schema_valid("checkpoint", checkpoint)
     assert checkpoint["route"] == "large_high_risk"
     assert checkpoint["current_stage"] == "clarify"
     assert checkpoint["next_action"] == "Resume at plan after reloading canonical artifacts."
